+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### 🚀 Features
 - Create product suggest endpoint ([#257](https://github.com/whirli/whirli-client/pull/257))
+### 🐛 Bug fixes
+- Refactored users resource path ([#256](https://github.com/whirli/whirli-client/pull/256))
 
 ## [0.0.36] - 2020-08-12
 ### 🚀 Features